+++ spec/lib/glimmer__list_data_binding__spec.rb
@@ -32,11 +32,11 @@ module GlimmerSpec
 
       class ::RedList
         include Glimmer::UI::CustomWidget
-        def body
+        body {
           list(swt_style) {
             background :red
           }
-        end
+        }
       end
     end